// src/app/models/user.ts
// Interface for a User object
export interface User {
  username: string;

  firstName: string;

  lastName: string;

  gender: string;

  age: number;

  dateOfBirth: Date;

  lastActive: Date;

  introduction: string;

  lookingFor: string;

  interests: string;

  city: string;

  country: string;

  mainPhoto: string;
}